fixes from suggestions
