+++ synapse/rest/client/v2_alpha/account.py
@@ -69,7 +69,7 @@ class EmailPasswordRequestTokenRestServlet(RestServlet):
     @defer.inlineCallbacks
     def on_POST(self, request):
         if self.config.threepid_behaviour == ThreepidBehaviour.OFF:
-            if self.config.local_threepid_emails_disabled_due_to_config:
+            if self.config.local_threepid_handling_disabled_due_to_email_config:
                 logger.warn(
                     "User password resets have been disabled due to lack of email config"
                 )
@@ -121,7 +121,7 @@ class EmailPasswordRequestTokenRestServlet(RestServlet):
             )
         elif self.config.threepid_behaviour == ThreepidBehaviour.LOCAL:
             # Send password reset emails from Synapse
-            sid = yield self.mailer.send_threepid_validation(
+            sid = yield self.identity_handler.send_threepid_validation(
                 email,
                 client_secret,
                 send_attempt,
@@ -136,77 +136,7 @@ class EmailPasswordRequestTokenRestServlet(RestServlet):
                 400, "Password reset by email is not supported on this homeserver"
             )
 
-        return (200, ret)
-
-    @defer.inlineCallbacks
-    def send_password_reset(self, email, client_secret, send_attempt, next_link=None):
-        """Send a password reset email
-
-        Args:
-            email (str): The user's email address
-            client_secret (str): The provided client secret
-            send_attempt (int): Which send attempt this is
-            next_link (str|None): The link to redirect the user to upon success. No redirect
-                occurs if None
-
-        Returns:
-            The new session_id upon success
-
-        Raises:
-            SynapseError is an error occurred when sending the email
-        """
-        # Check that this email/client_secret/send_attempt combo is new or
-        # greater than what we've seen previously
-        session = yield self.datastore.get_threepid_validation_session(
-            "email", client_secret, address=email, validated=False
-        )
-
-        # Check to see if a session already exists and that it is not yet
-        # marked as validated
-        if session and session.get("validated_at") is None:
-            session_id = session["session_id"]
-            last_send_attempt = session["last_send_attempt"]
-
-            # Check that the send_attempt is higher than previous attempts
-            if send_attempt <= last_send_attempt:
-                # If not, just return a success without sending an email
-                return session_id
-        else:
-            # An non-validated session does not exist yet.
-            # Generate a session id
-            session_id = random_string(16)
-
-        # Generate a new validation token
-        token = random_string(32)
-
-        # Send the mail with the link containing the token, client_secret
-        # and session_id
-        try:
-            yield self.mailer.send_password_reset_mail(
-                email, token, client_secret, session_id
-            )
-        except Exception:
-            logger.exception("Error sending a password reset email to %s", email)
-            raise SynapseError(
-                500, "An error was encountered when sending the password reset email"
-            )
-
-        token_expires = (
-            self.hs.clock.time_msec() + self.config.email_validation_token_lifetime
-        )
-
-        yield self.datastore.start_or_continue_validation_session(
-            "email",
-            email,
-            session_id,
-            client_secret,
-            send_attempt,
-            next_link,
-            token,
-            token_expires,
-        )
-
-        return session_id
+        return 200, ret
 
 
 class MsisdnPasswordRequestTokenRestServlet(RestServlet):
@@ -299,7 +229,7 @@ class PasswordResetSubmitTokenServlet(RestServlet):
                 400, "This medium is currently not supported for password resets"
             )
         if self.config.threepid_behaviour == ThreepidBehaviour.OFF:
-            if self.config.local_threepid_emails_disabled_due_to_config:
+            if self.config.local_threepid_handling_disabled_due_to_email_config:
                 logger.warn(
                     "Password reset emails have been disabled due to lack of an email config"
                 )
@@ -307,9 +237,9 @@ class PasswordResetSubmitTokenServlet(RestServlet):
                 400, "Email-based password resets are disabled on this server"
             )
 
-        sid = parse_string(request, "sid")
-        client_secret = parse_string(request, "client_secret")
-        token = parse_string(request, "token")
+        sid = parse_string(request, "sid", required=True)
+        client_secret = parse_string(request, "client_secret", required=True)
+        token = parse_string(request, "token", required=True)
 
         # Attempt to validate a 3PID session
         try:
@@ -344,7 +274,6 @@ class PasswordResetSubmitTokenServlet(RestServlet):
 
         request.write(html.encode("utf-8"))
         finish_request(request)
-        return None
 
     def load_jinja2_template(self, template_dir, template_filename, template_vars):
         """Loads a jinja2 template with variables to insert
@@ -367,7 +296,9 @@ class PasswordResetSubmitTokenServlet(RestServlet):
     @defer.inlineCallbacks
     def on_POST(self, request, medium):
         if medium != "email":
-            raise SynapseError(400, "This medium is currently not supported")
+            raise SynapseError(
+                400, "This medium is currently not supported for password resets"
+            )
 
         body = parse_json_object_from_request(request)
         assert_params_in_dict(body, ["sid", "client_secret", "token"])
@@ -504,20 +435,11 @@ class EmailThreepidRequestTokenRestServlet(RestServlet):
         super(EmailThreepidRequestTokenRestServlet, self).__init__()
         self.hs = hs
         self.config = hs.config
-        self.store = self.hs.get_datastore()
         self.identity_handler = hs.get_handlers().identity_handler
+        self.store = self.hs.get_datastore()
 
     @defer.inlineCallbacks
     def on_POST(self, request):
-        if self.config.threepid_behaviour == "off":
-            if self.config.local_threepid_emails_disabled_due_to_config:
-                logger.warn(
-                    "User password resets have been disabled due to lack of email config"
-                )
-            raise SynapseError(
-                400, "Adding email identities have been disabled on this server"
-            )
-
         body = parse_json_object_from_request(request)
         assert_params_in_dict(
             body, ["id_server", "client_secret", "email", "send_attempt"]
@@ -545,7 +467,7 @@ class EmailThreepidRequestTokenRestServlet(RestServlet):
         ret = yield self.identity_handler.requestEmailToken(
             id_server, email, client_secret, send_attempt, next_link
         )
-        return (200, ret)
+        return 200, ret
 
 
 class MsisdnThreepidRequestTokenRestServlet(RestServlet):
